+++ extensions/elation/__mocks__/constants.ts
@@ -1,4 +1,5 @@
 import { type AppointmentInput } from '../types/appointment'
+import { type PostLetterResponse } from '../types/letter'
 import { type NonVisitNoteResponse } from '../types/nonVisitNote'
 import { type PatientInput } from '../types/patient'
 import { type PhysicianResponse } from '../types/physician'
@@ -82,4 +83,17 @@ export const nonVisitNoteResponseExample: NonVisitNoteResponse = {
   chart_date: '2023-01-01T00:00:00Z',
   document_date: '2023-01-01T00:00:00Z',
   tags: [],
+}
+
+export const postLetterResponseExample: PostLetterResponse = {
+  body: 'This my test',
+  id: 142251583930586,
+  letter_type: 'provider',
+  patient: 141986488057857,
+  practice: 141127173275652,
+  referral_order: null,
+  send_to_contact: {
+    id: 141377681883138,
+  },
+  subject: 'My test subject',
 }